added router with some pages.

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,14 @@ import { RequestHandler } from 'express';
 import NoteModel from '../models/note';
 import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
+import { assertIsDefined } from '../util/assertisDefined';
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserID = req.session.userID;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserID);
+    const notes = await NoteModel.find({ userID: authenticatedUserID }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -14,15 +18,25 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteID = req.params.noteID;
+  const authenticatedUserID = req.session.userID;
+
   try {
+    assertIsDefined(authenticatedUserID);
+
     if (!mongoose.isValidObjectId(noteID)) {
       throw createHttpError(400, 'Invalid note ID');
     }
 
     const note = await NoteModel.findById(noteID).exec();
+
     if (!note) {
       throw createHttpError(404, 'note not found');
     }
+
+    if (!note.userID.equals(authenticatedUserID)) {
+      throw createHttpError(401, 'you cannot access this note!');
+    }
+
     res.status(200).json(note);
   } catch (error) {
     next(error);
@@ -34,18 +48,28 @@ interface CreateNoteBody {
   text?: string;
 }
 
-export const createNotes: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
+export const createNotes: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (
+  req,
+  res,
+  next
+) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authenticatedUserID = req.session.userID;
 
   try {
+    assertIsDefined(authenticatedUserID);
+
     if (!title) {
       throw createHttpError(400, 'Note must have a title');
     }
+
     const newNote = await NoteModel.create({
+      userID: authenticatedUserID,
       title,
       text,
     });
+
     res.status(201).json(newNote);
   } catch (error) {
     next(error);
@@ -61,12 +85,20 @@ interface UpdateNoteBody {
   text?: string;
 }
 
-export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody, unknown> = async (req, res, next) => {
+export const updateNote: RequestHandler<
+  UpdateNoteParams,
+  unknown,
+  UpdateNoteBody,
+  unknown
+> = async (req, res, next) => {
   const noteID = req.params.noteID;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserID = req.session.userID;
 
   try {
+    assertIsDefined(authenticatedUserID);
+
     if (!mongoose.isValidObjectId(noteID)) {
       throw createHttpError(400, 'Invalid note ID');
     }
@@ -78,6 +110,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const note = await NoteModel.findById(noteID).exec();
     if (!note) {
       throw createHttpError(404, 'note not found');
+    }
+
+    if (!note.userID.equals(authenticatedUserID)) {
+      throw createHttpError(401, 'you cannot access this note!');
     }
 
     note.title = newTitle;
@@ -92,7 +128,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteID = req.params.noteID;
+  const authenticatedUserID = req.session.userID;
   try {
+    assertIsDefined(authenticatedUserID);
+
     if (!mongoose.isValidObjectId(noteID)) {
       throw createHttpError(400, 'Invalid note ID');
     }
@@ -101,6 +140,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, 'Note not found');
+    }
+
+    if (!note.userID.equals(authenticatedUserID)) {
+      throw createHttpError(401, 'you cannot access this note!');
     }
 
     await NoteModel.findByIdAndDelete(note);
